backend/requirements.txt
flask
flask_cors
python-dotenv
flask_bcrypt
flask_jwt_extended
flask_mail